feat: store에서 JWT토큰을 이용하도록 변경 (2)

+++ src/stores/stores.controller.ts
@@ -38,7 +38,7 @@ export class StoresController {
         @Body() createEventDTO: CreateEventDTO
     ): Promise<ApiResponseDTO<void>> {
         const foundStore = await this.storesService.readStoreById(store_id)
-        if (foundStore.user_id.user_id !== req.user.user_id) {
+        if (foundStore.user.user_id !== req.user.user_id) {
             throw new ForbiddenException('You Can Create Event Your Own Store.')
         }
 
@@ -87,12 +87,12 @@ export class StoresController {
         @Body() updateEventDTO: UpdateEventDTO
     ): Promise<ApiResponseDTO<void>> {
         const foundStore = await this.storesService.readStoreById(store_id)
-        if (foundStore.user_id.user_id !== req.user.user_id) {
+        if (foundStore.user.user_id !== req.user.user_id) {
             throw new ForbiddenException('You Can Only Update Your Own Store Event.')
         }
 
         const foundEvent = await this.storesService.readEventById(event_id)
-        if (foundEvent.store.user_id.user_id !== req.user.user_id) {
+        if (foundEvent.store.user.user_id !== req.user.user_id) {
             throw new ForbiddenException('This Event Does Not Belong to the Provided Store.')
         }
 
@@ -112,12 +112,12 @@ export class StoresController {
         @Param('event_id') event_id: number
     ): Promise<ApiResponseDTO<void>> {
         const foundStore = await this.storesService.readStoreById(store_id)
-        if (foundStore.user_id.user_id !== req.user.user_id) {
+        if (foundStore.user.user_id !== req.user.user_id) {
             throw new ForbiddenException('You Can Only Delete Your Own Store Event.')
         }
 
         const foundEvent = await this.storesService.readEventById(event_id)
-        if (foundEvent.store.user_id.user_id !== req.user.user_id) {
+        if (foundEvent.store.user.user_id !== req.user.user_id) {
             throw new ForbiddenException('This Event Does Not Belong to the Provided Store.')
         }
 
@@ -131,16 +131,23 @@ export class StoresController {
     // CREATE
     // 새로운 가게 생성하기
     @Post('/')
-    async createStore(@Body() createStoreDTO: CreateStoreDTO): Promise<ApiResponseDTO<void>> {
-        await this.storesService.createStore(createStoreDTO)
+    @Roles(UserRole.MANAGER)
+    async createStore(
+        @Req() req: AuthenticatedRequest,
+        @Body() createStoreDTO: CreateStoreDTO
+    ): Promise<ApiResponseDTO<void>> {
+        const user_id = req.user.user_id
+        await this.storesService.createStore(user_id, createStoreDTO)
 
         return new ApiResponseDTO(true, HttpStatus.CREATED, "Store Created Successfully")
     }
 
     // READ - 매니저가 자신의 가게 조회할 때 사용
-    @Get('/user/:user_id')
+    @Get('/my')
     @Roles(UserRole.MANAGER)
-    async readAllStoresByUser(@Param('user_id') user_id: number): Promise<ApiResponseDTO<ReadStoreDTO[]>> {
+    async readAllStoresByUser(@Req() req: AuthenticatedRequest): Promise<ApiResponseDTO<ReadStoreDTO[]>> {
+        const user_id = req.user.user_id
+
         const stores = await this.storesService.readAllStoresByUser(user_id)
         const readStoreDTOs = stores.map(store => new ReadStoreDTO(store))
 
@@ -186,7 +193,10 @@ export class StoresController {
     // 가게 매니저 수정 (관리자 전용)
     @Patch('/:store_id')
     @Roles(UserRole.ADMIN)
-    async updateStoreManager(@Param('store_id') store_id: number, @Body('user_id') user_id: number): Promise<ApiResponseDTO<void>> {
+    async updateStoreManager(
+        @Param('store_id') store_id: number,
+        @Body('user_id') user_id: number
+    ): Promise<ApiResponseDTO<void>> {
         await this.storesService.updateStoreManager(store_id, user_id)
 
         return new ApiResponseDTO(true, HttpStatus.OK, "Store Manager Updated Successfully")
@@ -194,8 +204,17 @@ export class StoresController {
 
     // 가게 정보 수정 (매니저 전용)
     @Put('/:store_id')
-    @Roles(UserRole.MANAGER)
-    async updateStoreDetail(@Param('store_id') store_id: number, @Body() updateStoreDetailDTO: UpdateStoreDetailDTO): Promise<ApiResponseDTO<void>> {
+    @Roles(UserRole.MANAGER, UserRole.ADMIN)
+    async updateStoreDetail(
+        @Req() req: AuthenticatedRequest,
+        @Param('store_id') store_id: number,
+        @Body() updateStoreDetailDTO: UpdateStoreDetailDTO
+    ): Promise<ApiResponseDTO<void>> {
+        const foundStore = await this.storesService.readStoreById(store_id)
+        if (foundStore.user.user_id !== req.user.user_id) {
+            throw new ForbiddenException('You Can Create Store Info Your Own Store.')
+        }
+
         await this.storesService.updateStoreDetail(store_id, updateStoreDetailDTO)
 
         return new ApiResponseDTO(true, HttpStatus.OK, "Store Information Updated Successfully")
@@ -204,7 +223,15 @@ export class StoresController {
     // DELETE
     @Delete('/:store_id')
     @Roles(UserRole.ADMIN, UserRole.MANAGER)
-    async deleteStore(@Param('store_id') store_id: number): Promise<ApiResponseDTO<void>> {
+    async deleteStore(
+        @Req() req: AuthenticatedRequest,
+        @Param('store_id') store_id: number
+    ): Promise<ApiResponseDTO<void>> {
+        const foundStore = await this.storesService.readStoreById(store_id)
+        if (foundStore.user.user_id !== req.user.user_id) {
+            throw new ForbiddenException('You Can Delete Store Info Your Own Store.')
+        }
+
         await this.storesService.deleteStore(store_id)
 
         return new ApiResponseDTO(true, HttpStatus.OK, "Store Deleted Successfully")
@@ -214,7 +241,10 @@ export class StoresController {
     // 새로운 리뷰 생성
     @Post('/:store_id/reviews')
     @Roles(UserRole.USER)
-    async createReview(@Param('store_id') store_id: number, @Body() createReviewDTO: CreateReviewDTO): Promise<ApiResponseDTO<void>> {
+    async createReview(
+        @Param('store_id') store_id: number,
+        @Body() createReviewDTO: CreateReviewDTO
+    ): Promise<ApiResponseDTO<void>> {
         await this.reviewsService.createReview(store_id, createReviewDTO)
 
         return new ApiResponseDTO(true, HttpStatus.CREATED, 'Review Created Successfully!')
